--- practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/GroupDeletionTest.java
@@ -1,14 +1,0 @@
-package ru.stqa.ptf.addressbook;
-
-import org.testng.annotations.Test;
-
-public class GroupDeletionTest extends TestBase {
-
-    @Test
-    public void testGroupDeletion() throws Exception {
-        goToGroupPage();
-        selectGroup();
-        deleteSelectedGroup();
-        returnToGroupPage();
-    }
-}
