+++ practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/appmanager/ApplicationManager.java
@@ -1,0 +1,81 @@
+package ru.stqa.ptf.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.NoAlertPresentException;
+import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
+
+public class ApplicationManager {
+    public WebDriver wd;
+
+    private SessionHelper sessionHelper;
+    private NavigationHelper navigationHelper;
+    private ContactHelper contactHelper;
+    private GroupHelper groupHelper;
+
+    public WebDriverWait wait;
+
+    public void init() {
+        System.setProperty("webdriver.chrome.driver", "chromedriver.exe");
+        wd = new ChromeDriver();
+        wd.manage().timeouts().implicitlyWait(Duration.ofSeconds(30));
+        wait = new WebDriverWait(wd, Duration.ofSeconds(10));
+        wd.get("http://127.0.0.1/addressbook/");
+        navigationHelper = new NavigationHelper(wd);
+        groupHelper = new GroupHelper(wd);
+        contactHelper = new ContactHelper(wd);
+        sessionHelper = new SessionHelper(wd);
+
+        sessionHelper.login("admin", "secret");
+    }
+
+    public void goToHomePage() {
+        wd.findElement(By.linkText("home page")).click();
+    }
+
+    public void goToAddNewContact() {
+        wd.findElement(By.linkText("add new")).click();
+    }
+
+    public void logout() {
+        wd.findElement(By.linkText("Logout")).click();
+    }
+
+    public void stop() {
+        wd.quit();
+    }
+
+    public boolean isElementPresent(By by) {
+        try {
+            wd.findElement(by);
+            return true;
+        } catch (NoSuchElementException e) {
+            return false;
+        }
+    }
+
+    public boolean isAlertPresent() {
+        try {
+            wd.switchTo().alert();
+            return true;
+        } catch (NoAlertPresentException e) {
+            return false;
+        }
+    }
+
+    public GroupHelper getGroupHelper() {
+        return groupHelper;
+    }
+
+    public NavigationHelper getNavigationHelper() {
+        return navigationHelper;
+    }
+
+    public ContactHelper getContactHelper() {
+        return contactHelper;
+    }
+}
