+++ practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/appmanager/ContactHelper.java
@@ -1,0 +1,61 @@
+package ru.stqa.ptf.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.Select;
+import ru.stqa.ptf.addressbook.model.ContactsData;
+
+public class ContactHelper {
+    private WebDriver wd;
+
+    public ContactHelper(WebDriver wd) {
+        this.wd = wd;
+    }
+
+    public void fillFormContact(ContactsData contactsData) {
+        wd.findElement(By.name("firstname")).click();
+        wd.findElement(By.name("firstname")).clear();
+        wd.findElement(By.name("firstname")).sendKeys(contactsData.getName());
+        wd.findElement(By.name("middlename")).click();
+        wd.findElement(By.name("middlename")).clear();
+        wd.findElement(By.name("middlename")).sendKeys(contactsData.getMiddleName());
+        wd.findElement(By.name("lastname")).click();
+        wd.findElement(By.name("lastname")).clear();
+        wd.findElement(By.name("lastname")).sendKeys(contactsData.getLastName());
+        wd.findElement(By.name("nickname")).click();
+        wd.findElement(By.name("nickname")).clear();
+        wd.findElement(By.name("nickname")).sendKeys(contactsData.getNickName());
+        wd.findElement(By.name("company")).click();
+        wd.findElement(By.name("company")).clear();
+        wd.findElement(By.name("company")).sendKeys(contactsData.getCompany());
+        wd.findElement(By.name("address")).click();
+        wd.findElement(By.name("address")).clear();
+        wd.findElement(By.name("address")).sendKeys(contactsData.getAddress());
+        wd.findElement(By.name("home")).click();
+        wd.findElement(By.name("home")).clear();
+        wd.findElement(By.name("home")).sendKeys(contactsData.getHomePhoneNumber());
+        wd.findElement(By.name("mobile")).click();
+        wd.findElement(By.name("mobile")).clear();
+        wd.findElement(By.name("mobile")).sendKeys(contactsData.getMobilePhoneNumber());
+        wd.findElement(By.name("work")).click();
+        wd.findElement(By.name("work")).clear();
+        wd.findElement(By.name("work")).sendKeys(contactsData.getWork());
+        wd.findElement(By.name("fax")).click();
+        wd.findElement(By.name("fax")).clear();
+        wd.findElement(By.name("fax")).sendKeys(contactsData.getFax());
+        wd.findElement(By.name("email")).click();
+        wd.findElement(By.name("email")).clear();
+        wd.findElement(By.name("email")).sendKeys(contactsData.getEmail());
+        wd.findElement(By.name("bday")).click();
+        new Select(wd.findElement(By.name("bday"))).selectByVisibleText(contactsData.getBirthDay());
+        wd.findElement(By.name("theform")).click();
+        new Select(wd.findElement(By.name("bmonth"))).selectByVisibleText(contactsData.getBirthMonth());
+        wd.findElement(By.name("byear")).click();
+        wd.findElement(By.name("byear")).clear();
+        wd.findElement(By.name("byear")).sendKeys(contactsData.getBirthYear());
+        wd.findElement(By.name("new_group")).click();
+        new Select(wd.findElement(By.name("new_group"))).selectByVisibleText(contactsData.getGroup());
+        wd.findElement(By.xpath("//div[@id='content']/form/input[21]")).click();
+
+    }
+}
