+++ practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/appmanager/NavigationHelper.java
@@ -1,0 +1,16 @@
+package ru.stqa.ptf.addressbook.appmanager;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+
+public class NavigationHelper {
+    private WebDriver wd;
+
+    public NavigationHelper(WebDriver wd) {
+        this.wd = wd;
+    }
+
+    public void goToGroupPage() {
+        wd.findElement(By.linkText("groups")).click();
+    }
+}
