+++ practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/model/ContactsData.java
@@ -1,4 +1,4 @@
-package ru.stqa.ptf.addressbook;
+package ru.stqa.ptf.addressbook.model;
 
 public class ContactsData {
     private final String name;
