+++ practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/tests/GroupDeletionTest.java
@@ -1,0 +1,14 @@
+package ru.stqa.ptf.addressbook.tests;
+
+import org.testng.annotations.Test;
+
+public class GroupDeletionTest extends TestBase {
+
+    @Test
+    public void testGroupDeletion() throws Exception {
+        app.getNavigationHelper().goToGroupPage();
+        app.getGroupHelper().selectGroup();
+        app.getGroupHelper().deleteSelectedGroup();
+        app.getGroupHelper().returnToGroupPage();
+    }
+}
