+++ practiceMakesPerfectAqaJava/addressbook-web-tests/src/test/java/ru/stqa/ptf/addressbook/tests/TestBase.java
@@ -1,0 +1,20 @@
+package ru.stqa.ptf.addressbook.tests;
+
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+import ru.stqa.ptf.addressbook.appmanager.ApplicationManager;
+
+public class TestBase {
+
+    protected final ApplicationManager app = new ApplicationManager();
+
+    @BeforeMethod
+    public void setUp() throws Exception {
+        app.init();
+    }
+
+    @AfterMethod
+    public void tearDown() throws Exception {
+        app.stop();
+    }
+}
