--- practiceMakesPerfectAqaJava/addressbook-web-tests/tests.uml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Diagram>
-  <ID>JAVA</ID>
-  <OriginalElement>C:/D/myEducation/java/javaProgrammingForTesters/practiceMakesPerfectAqaJava/practiceMakesPerfectAqaJava/addressbook-web-tests</OriginalElement>
-  <nodes>
-    <node x="-93.0" y="226.0">ru.stqa.ptf.addressbook.GroupCreationTest</node>
-    <node x="28.0" y="166.0">ru.stqa.ptf.addressbook.GroupDeletionTest</node>
-    <node x="29.0" y="-14.0">ru.stqa.ptf.addressbook.ContactsData</node>
-    <node x="-55.0" y="46.0">ru.stqa.ptf.addressbook.TestBase</node>
-  </nodes>
-  <notes />
-  <edges>
-    <edge source="ru.stqa.ptf.addressbook.GroupDeletionTest" target="ru.stqa.ptf.addressbook.TestBase" relationship="GENERALIZATION">
-      <point x="0.0" y="-14.0" />
-      <point x="126.0" y="140.0" />
-      <point x="6.0" y="140.0" />
-      <point x="0.0" y="14.0" />
-    </edge>
-    <edge source="ru.stqa.ptf.addressbook.GroupCreationTest" target="ru.stqa.ptf.addressbook.TestBase" relationship="GENERALIZATION">
-      <point x="0.0" y="-14.0" />
-      <point x="0.0" y="14.0" />
-    </edge>
-  </edges>
-  <settings layout="Hierarchic Compact" zoom="1.5" showDependencies="false" x="75.5" y="130.0" />
-  <SelectedNodes>
-    <node>ru.stqa.ptf.addressbook.ContactsData</node>
-  </SelectedNodes>
-  <Categories />
-  <SCOPE>All</SCOPE>
-  <VISIBILITY>private</VISIBILITY>
-</Diagram>
-
